# Improved HTTP/1 header parser