… src/libraries/System.Net.Http/src/System/Net/Http/SocketsHttpHandler/HttpConnection.cs