This replaces the SocketsHttpHandler header parser with two implementations: AVX2, and portable. This ports over the work done as part of the LLHTTP project.

This gives a 5-10% end-to-end perf improvement over localhost:

| Faster                                                                           | base/diff | Base Median (ns) | Diff Median (ns) | Modality|
| -------------------------------------------------------------------------------- | ---------:| ----------------:| ----------------:| -------- |
| System.Net.Http.Tests.SocketsHttpHandlerPerfTest.Get(ssl: False, chunkedResponse |      1.09 |         34244.28 |         31391.45 |         |
| System.Net.Http.Tests.SocketsHttpHandlerPerfTest.Get(ssl: False, chunkedResponse |      1.08 |         33410.35 |         30836.15 |         |
| System.Net.Http.Tests.SocketsHttpHandlerPerfTest.Get_EnumerateHeaders_Validated( |      1.08 |         34049.97 |         31524.56 |         |
| System.Net.Http.Tests.SocketsHttpHandlerPerfTest.Get_EnumerateHeaders_Unvalidate |      1.07 |         37427.29 |         34867.46 | several?|

The AVX2 parser has 50-100% performance improvement over the portable implementation. It is faster in every situation (few headers, lots of headers, small headers, large headers).

|          Method | HeaderCount | HeaderNameSize | HeaderValueSize |      Mean |    Error |   StdDev |
|---------------- |------------ |--------------- |---------------- |----------:|---------:|---------:|
| PortableHeaders |           1 |              1 |               1 |  14.20 ns | 0.078 ns | 0.073 ns |
|     Avx2Headers |           1 |              1 |               1 |  10.50 ns | 0.094 ns | 0.088 ns |
| PortableHeaders |           1 |              1 |               8 |  15.06 ns | 0.096 ns | 0.090 ns |
|     Avx2Headers |           1 |              1 |               8 |  10.48 ns | 0.069 ns | 0.065 ns |
| PortableHeaders |           1 |              1 |             128 |  18.05 ns | 0.089 ns | 0.084 ns |
|     Avx2Headers |           1 |              1 |             128 |  12.20 ns | 0.112 ns | 0.105 ns |
| PortableHeaders |           1 |              8 |               1 |  14.26 ns | 0.087 ns | 0.081 ns |
|     Avx2Headers |           1 |              8 |               1 |  10.49 ns | 0.080 ns | 0.071 ns |
| PortableHeaders |           1 |              8 |               8 |  14.83 ns | 0.074 ns | 0.065 ns |
|     Avx2Headers |           1 |              8 |               8 |  10.52 ns | 0.073 ns | 0.069 ns |
| PortableHeaders |           1 |              8 |             128 |  16.38 ns | 0.103 ns | 0.091 ns |
|     Avx2Headers |           1 |              8 |             128 |  12.73 ns | 0.075 ns | 0.066 ns |
| PortableHeaders |           1 |             16 |               1 |  14.15 ns | 0.129 ns | 0.108 ns |
|     Avx2Headers |           1 |             16 |               1 |  10.67 ns | 0.115 ns | 0.096 ns |
| PortableHeaders |           1 |             16 |               8 |  15.27 ns | 0.227 ns | 0.201 ns |
|     Avx2Headers |           1 |             16 |               8 |  10.69 ns | 0.115 ns | 0.108 ns |
| PortableHeaders |           1 |             16 |             128 |  18.44 ns | 0.143 ns | 0.133 ns |
|     Avx2Headers |           1 |             16 |             128 |  13.01 ns | 0.243 ns | 0.227 ns |
| PortableHeaders |           8 |              1 |               1 |  56.03 ns | 0.711 ns | 0.665 ns |
|     Avx2Headers |           8 |              1 |               1 |  29.64 ns | 0.430 ns | 0.403 ns |
| PortableHeaders |           8 |              1 |               8 |  67.30 ns | 0.311 ns | 0.275 ns |
|     Avx2Headers |           8 |              1 |               8 |  31.10 ns | 0.180 ns | 0.151 ns |
| PortableHeaders |           8 |              1 |             128 |  82.21 ns | 0.448 ns | 0.397 ns |
|     Avx2Headers |           8 |              1 |             128 |  50.41 ns | 0.593 ns | 0.555 ns |
| PortableHeaders |           8 |              8 |               1 |  58.42 ns | 1.005 ns | 0.940 ns |
|     Avx2Headers |           8 |              8 |               1 |  30.94 ns | 0.189 ns | 0.176 ns |
| PortableHeaders |           8 |              8 |               8 |  72.81 ns | 0.912 ns | 0.853 ns |
|     Avx2Headers |           8 |              8 |               8 |  31.56 ns | 0.133 ns | 0.124 ns |
| PortableHeaders |           8 |              8 |             128 |  86.82 ns | 0.714 ns | 0.633 ns |
|     Avx2Headers |           8 |              8 |             128 |  51.55 ns | 0.286 ns | 0.267 ns |
| PortableHeaders |           8 |             16 |               1 |  55.19 ns | 0.160 ns | 0.149 ns |
|     Avx2Headers |           8 |             16 |               1 |  32.95 ns | 0.297 ns | 0.278 ns |
| PortableHeaders |           8 |             16 |               8 |  74.59 ns | 0.375 ns | 0.351 ns |
|     Avx2Headers |           8 |             16 |               8 |  34.22 ns | 0.121 ns | 0.108 ns |
| PortableHeaders |           8 |             16 |             128 |  84.59 ns | 0.311 ns | 0.276 ns |
|     Avx2Headers |           8 |             16 |             128 |  51.96 ns | 0.298 ns | 0.279 ns |
| PortableHeaders |          16 |              1 |               1 | 119.45 ns | 0.736 ns | 0.688 ns |
|     Avx2Headers |          16 |              1 |               1 |  52.34 ns | 0.367 ns | 0.306 ns |
| PortableHeaders |          16 |              1 |               8 | 134.94 ns | 0.466 ns | 0.436 ns |
|     Avx2Headers |          16 |              1 |               8 |  53.98 ns | 0.322 ns | 0.301 ns |
| PortableHeaders |          16 |              1 |             128 | 171.17 ns | 1.379 ns | 1.290 ns |
|     Avx2Headers |          16 |              1 |             128 | 100.83 ns | 0.502 ns | 0.445 ns |
| PortableHeaders |          16 |              8 |               1 | 109.07 ns | 0.498 ns | 0.465 ns |
|     Avx2Headers |          16 |              8 |               1 |  53.92 ns | 0.129 ns | 0.120 ns |
| PortableHeaders |          16 |              8 |               8 | 140.69 ns | 0.591 ns | 0.553 ns |
|     Avx2Headers |          16 |              8 |               8 |  56.92 ns | 0.243 ns | 0.203 ns |
| PortableHeaders |          16 |              8 |             128 | 169.72 ns | 0.592 ns | 0.494 ns |
|     Avx2Headers |          16 |              8 |             128 |  99.44 ns | 0.401 ns | 0.375 ns |
| PortableHeaders |          16 |             16 |               1 | 114.38 ns | 0.338 ns | 0.282 ns |
|     Avx2Headers |          16 |             16 |               1 |  56.66 ns | 0.152 ns | 0.142 ns |
| PortableHeaders |          16 |             16 |               8 | 142.28 ns | 0.481 ns | 0.402 ns |
|     Avx2Headers |          16 |             16 |               8 |  60.48 ns | 0.328 ns | 0.306 ns |
| PortableHeaders |          16 |             16 |             128 | 167.60 ns | 0.416 ns | 0.389 ns |
|     Avx2Headers |          16 |             16 |             128 | 108.95 ns | 0.446 ns | 0.417 ns |