@scalablecory 2022-01-03
> The portable version uses span.IndexOf and span.IndexOfAny. My understanding is that these are already vector-optimized. Is that not happening here for some reason?
> 
> Or, is the hand-rolled vectorization more efficient for some reason, e.g. we are amortizing some costs across operations or something like that?

Those are vectorized, but they throw away a lot of data that is useful for optimizing HTTP/1. In our case, 32 bytes will likely contain 2-3 headers, and this maintains (in a bitmask) all the interesting tokens to check. If we assume 3 headers, using `IndexOf` means 6 SIMD scans (for `:` and `\n`), while this would do just 1.

There is also perf improvement by removing bounds checks and hand optimizing for good codegen, but this was secondary.