@geoffkizer 2022-01-03
I wonder if it would be better to have "intro" and "outro" code to handle misaligned prefix/postfix data.

This seems like it has a couple potential benefits:
(1) Don't need SimdPaddedArray at all
(2) Seems like it would remove some special cases in the inner loop, e.g. artificial match prior to beginning or after end of buffer
(3) Code is more common between different cases, meaning no-vector/vector128/vector256avx

Thoughts?