@scalablecory 2022-01-03
> I wonder if it would be better to have "intro" and "outro" code to handle misaligned prefix/postfix data.
> 
> This seems like it has a couple potential benefits: (1) Don't need SimdPaddedArray at all (2) Seems like it would remove some special cases in the inner loop, e.g. artificial match prior to beginning or after end of buffer (3) Code is more common between different cases, meaning no-vector/vector128/vector256avx
> 
> Thoughts?

Intro/outro approach is a no-go -- I had a version that did this and it was much slower. I also had a version that used a load mask which was also slower.

In theory the code should never match past the buffer because that will be 0-initialized; let me check that this is the case. I can't remember why I put the check for that back in.

We could use unaligned loads, or pad the front of the buffer with 0s too, to potentially avoid some checks. What I found was that matches in the "pre-buffer" area were rare enough that I wasn't sure the tradeoff of a mostly perfectly predicted branch VS the setup to remove the branch was worth it. I changed the code a bit since I tried this, though, so maybe it's worth another look.