@stephentoub 2022-01-03
> Those are vectorized, but they throw away a lot of data that is useful for optimizing HTTP/1. In our case, 32 bytes will likely contain 2-3 headers, and this maintains (in a bitmask) all the interesting tokens to check. If we assume 3 headers, using IndexOf means 6 SIMD scans (for : and \n), while this would do just 1.

Can you elaborate on this?  Specifically:
1. What is the data source for 32 bytes containing 2-3 headers?  Just a non-scientific test, when I do a get request to bing.com, the average header length is 150 chars, for httpbin.org it's 31, etc.
2. What do you mean about IndexOf doing 6 scans whereas this doing only 1?  IndexOfAny(char1, char2) similarly ors together a mask for each character; did you mean IndexOfAny doing 3 vs this doing only 1?