@scalablecory 2022-01-03
> > Based on prior experience working with a wide variety of web APIs. Websites certainly tend to have much larger headers, which is why I benchmarked against that as well.
> 
> Even the minimalist default web forecast microservice ASP.NET template has an average length of 27. I'm not discounting prior experience, but if we're going to base decisions on this, it'd be helpful to see the concrete data impacting it.

Treat my anecdote as a scenario to show where `IndexOf` can be sub-optimal. It was on my mind because it was the inspiration for trying this out. It wasn't intended to distract or be stated as a reason to merge this PR -- we can have benchmarks for that. If you'd like to see more benchmarks, I can do that.