@stephentoub 2022-01-04
> It wasn't intended to distract

I'm not distracted ;-)  I'm trying to better understand the root of a cited ~35% improvement.  The example claimed the bulk of that improvement came from an ~6:1 reduction in calls to IndexOf{Any}.  It sounds instead like in the typical case it might really be more like 2:1, essentially that we're saving on duplicated overheads in setup costs from the IndexOfAny+IndexOf vs this effectively merging them together.  But there are other changes in the PR as well, so it's not clear to me how much is coming from that versus from other changes.  And if it really is all coming from the vectorization, that concerns me as well, in that there's a lot of complexity being added here, and intrinsics/vectorized code is orders of magnitude harder to maintain; while System.Net.Http is a little special, it's still representative of fairly typical library code, and I don't want the majority of such devs to need to replace an IndexOfAny+IndexOf simple code with hand-tuned vectorized code in order to gain an additional 35%.  Which then leads me to question if that really is the source of the gain, how can we do better?  Can we push any improvements down into IndexOfAny/IndexOf?  Can we add a new variant of these that can be used for key/value pair parsing and encapsulate this common pattern for SocketsHttpHandler to use but also for anyone else doing such key/value pair parsing to use (e.g. `Utf8Parser.TryParseKeyValuePair(byte separator, byte end)`)?  Etc.