@scalablecory 2022-01-04
I'm going to close this PR and split it up.

AVX is not the bulk of the improvement here, so the portable change will be a no-brainer.

I'll also open up an issue detailing the AVX parser to maybe spark some ideas on how we make parsing better in .NET.