@stephentoub 2022-01-04
> AVX is not the bulk of the improvement here, so the portable change will be a no-brainer.

Great. Thanks.